Move new handler to its own package.
