+++ db/user_test.go
@@ -42,70 +42,6 @@ func TestUserToFirestoreUpdate(t *testing.T) {
 	})
 }
 
-func TestGetUser(t *testing.T) {
-	d, err := Open(context.Background(), os.Getenv("TLACFG"))
-	if !assert.NoError(t, err) {
-		return
-	}
-
-	t.Run("MissingUID", func(t *testing.T) {
-		req := httptest.NewRequest(http.MethodGet, "/", nil)
-		rec := httptest.NewRecorder()
-		assert.NotNil(t, req, rec)
-		c := echo.New().NewContext(req, rec)
-
-		if assert.NoError(t, d.GetUser(c)) {
-			assert.Equal(t, http.StatusBadRequest, rec.Code)
-		}
-	})
-	t.Run("BadUID", func(t *testing.T) {
-		req := httptest.NewRequest(http.MethodGet, "/?uid=fakeUID", nil)
-		rec := httptest.NewRecorder()
-		assert.NotNil(t, req, rec)
-		c := echo.New().NewContext(req, rec)
-
-		if assert.NoError(t, d.GetUser(c)) {
-			assert.Equal(t, http.StatusNotFound, rec.Code)
-		}
-	})
-
-	dbConsistencyWarning(t)
-
-	t.Run("TypicalRequests", func(t *testing.T) {
-		// get some random doc
-		iter := d.Collection(usersPath).Documents(context.Background())
-		defer iter.Stop()
-		randomDoc, err := iter.Next()
-		assert.NoError(t, err)
-		t.Logf("using doc ID (%s)", randomDoc.Ref.ID)
-
-		t.Run("NoPrograms", func(t *testing.T) {
-			req := httptest.NewRequest(http.MethodGet, "/?uid="+randomDoc.Ref.ID, nil)
-			rec := httptest.NewRecorder()
-			assert.NotNil(t, req, rec)
-			c := echo.New().NewContext(req, rec)
-
-			if assert.NoError(t, d.GetUser(c)) {
-				assert.NotEmpty(t, rec.Result().Body)
-				assert.Equal(t, http.StatusOK, rec.Code)
-				// TODO: check for the rest of the response's validity
-			}
-		})
-		t.Run("WithPrograms", func(t *testing.T) {
-			req := httptest.NewRequest(http.MethodGet, "/?uid="+randomDoc.Ref.ID+"&programs=true", nil)
-			rec := httptest.NewRecorder()
-			assert.NotNil(t, req, rec)
-			c := echo.New().NewContext(req, rec)
-
-			if assert.NoError(t, d.GetUser(c)) {
-				assert.NotEmpty(t, rec.Result().Body)
-				assert.Equal(t, http.StatusOK, rec.Code)
-				// TODO: check for the rest of the response's validity
-			}
-		})
-	})
-}
-
 func TestUpdateUser(t *testing.T) {
 	d, err := Open(context.Background(), os.Getenv("TLACFG"))
 	if !assert.NoError(t, err) {
